404 return
Some of the cases, when a page is not found 404 will be added to result. For example, if you tried to grab FreeCompany with id 1, results will have ['freecompanies']['1'] = 404. This is done to easily identify cases of 404. If a different error was encountered, the entity will not be added to results.